# Convert CryptoToFiatConverter into a Singleton